planning: fix some float and double conversion warning.

+++ modules/planning/toolkits/optimizers/road_graph/trajectory_cost.cc
@@ -53,7 +53,7 @@ TrajectoryCost::TrajectoryCost(const DpPolyPathConfig &config,
       vehicle_param_(vehicle_param),
       heuristic_speed_data_(heuristic_speed_data),
       init_sl_point_(init_sl_point) {
-  const float total_time =
+  const double total_time =
       std::min(heuristic_speed_data_.TotalTime(), FLAGS_prediction_total_time);
 
   num_of_time_stamps_ = static_cast<uint32_t>(
@@ -67,9 +67,9 @@ TrajectoryCost::TrajectoryCost(const DpPolyPathConfig &config,
     }
     const auto &sl_boundary = ptr_obstacle->PerceptionSLBoundary();
 
-    const float adc_left_l =
+    const double adc_left_l =
         init_sl_point_.l() + vehicle_param_.left_edge_to_center();
-    const float adc_right_l =
+    const double adc_right_l =
         init_sl_point_.l() - vehicle_param_.right_edge_to_center();
 
     if (adc_left_l + FLAGS_lateral_ignore_buffer < sl_boundary.start_l() ||
@@ -96,7 +96,7 @@ TrajectoryCost::TrajectoryCost(const DpPolyPathConfig &config,
             ptr_obstacle->GetPointAtTime(t * config.eval_time_interval());
 
         Box2d obstacle_box = ptr_obstacle->GetBoundingBox(trajectory_point);
-        constexpr float kBuff = 0.5;
+        constexpr double kBuff = 0.5;
         Box2d expanded_obstacle_box =
             Box2d(obstacle_box.center(), obstacle_box.heading(),
                   obstacle_box.length() + kBuff, obstacle_box.width() + kBuff);
@@ -108,24 +108,24 @@ TrajectoryCost::TrajectoryCost(const DpPolyPathConfig &config,
 }
 
 ComparableCost TrajectoryCost::CalculatePathCost(
-    const QuinticPolynomialCurve1d &curve, const float start_s,
-    const float end_s, const uint32_t curr_level, const uint32_t total_level) {
+    const QuinticPolynomialCurve1d &curve, const double start_s,
+    const double end_s, const uint32_t curr_level, const uint32_t total_level) {
   ComparableCost cost;
-  float path_cost = 0.0;
-  std::function<float(const float)> quasi_softmax = [this](const float x) {
-    const float l0 = this->config_.path_l_cost_param_l0();
-    const float b = this->config_.path_l_cost_param_b();
-    const float k = this->config_.path_l_cost_param_k();
+  double path_cost = 0.0;
+  std::function<double(const double)> quasi_softmax = [this](const double x) {
+    const double l0 = this->config_.path_l_cost_param_l0();
+    const double b = this->config_.path_l_cost_param_b();
+    const double k = this->config_.path_l_cost_param_k();
     return (b + std::exp(-k * (x - l0))) / (1.0 + std::exp(-k * (x - l0)));
   };
 
   const auto &vehicle_config =
       common::VehicleConfigHelper::Instance()->GetConfig();
-  const float width = vehicle_config.vehicle_param().width();
+  const double width = vehicle_config.vehicle_param().width();
 
-  for (float curve_s = 0.0; curve_s < (end_s - start_s);
+  for (double curve_s = 0.0; curve_s < (end_s - start_s);
        curve_s += config_.path_resolution()) {
-    const float l = curve.Evaluate(0, curve_s);
+    const double l = curve.Evaluate(0, curve_s);
 
     path_cost += l * l * config_.path_l_cost() * quasi_softmax(std::fabs(l));
 
@@ -133,22 +133,22 @@ ComparableCost TrajectoryCost::CalculatePathCost(
     double right_width = 0.0;
     reference_line_->GetLaneWidth(curve_s + start_s, &left_width, &right_width);
 
-    constexpr float kBuff = 0.2;
+    constexpr double kBuff = 0.2;
     if (!is_change_lane_path_ && (l + width / 2.0 + kBuff > left_width ||
                                   l - width / 2.0 - kBuff < -right_width)) {
       cost.cost_items[ComparableCost::OUT_OF_BOUNDARY] = true;
     }
 
-    const float dl = std::fabs(curve.Evaluate(1, curve_s));
+    const double dl = std::fabs(curve.Evaluate(1, curve_s));
     path_cost += dl * dl * config_.path_dl_cost();
 
-    const float ddl = std::fabs(curve.Evaluate(2, curve_s));
+    const double ddl = std::fabs(curve.Evaluate(2, curve_s));
     path_cost += ddl * ddl * config_.path_ddl_cost();
   }
   path_cost *= config_.path_resolution();
 
   if (curr_level == total_level) {
-    const float end_l = curve.Evaluate(0, end_s - start_s);
+    const double end_l = curve.Evaluate(0, end_s - start_s);
     path_cost +=
         std::sqrt(end_l - init_sl_point_.l() / 2.0) * config_.path_end_l_cost();
   }
@@ -157,12 +157,12 @@ ComparableCost TrajectoryCost::CalculatePathCost(
 }
 
 ComparableCost TrajectoryCost::CalculateStaticObstacleCost(
-    const QuinticPolynomialCurve1d &curve, const float start_s,
-    const float end_s) {
+    const QuinticPolynomialCurve1d &curve, const double start_s,
+    const double end_s) {
   ComparableCost obstacle_cost;
-  for (float curr_s = start_s; curr_s <= end_s;
+  for (double curr_s = start_s; curr_s <= end_s;
        curr_s += config_.path_resolution()) {
-    const float curr_l = curve.Evaluate(0, curr_s - start_s);
+    const double curr_l = curve.Evaluate(0, curr_s - start_s);
     for (const auto &obs_sl_boundary : static_obstacle_sl_boundaries_) {
       obstacle_cost += GetCostFromObsSL(curr_s, curr_l, obs_sl_boundary);
     }
@@ -172,19 +172,19 @@ ComparableCost TrajectoryCost::CalculateStaticObstacleCost(
 }
 
 ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
-    const QuinticPolynomialCurve1d &curve, const float start_s,
-    const float end_s) const {
+    const QuinticPolynomialCurve1d &curve, const double start_s,
+    const double end_s) const {
   ComparableCost obstacle_cost;
   if (dynamic_obstacle_boxes_.empty()) {
     return obstacle_cost;
   }
 
-  float time_stamp = 0.0;
+  double time_stamp = 0.0;
   for (size_t index = 0; index < num_of_time_stamps_;
        ++index, time_stamp += config_.eval_time_interval()) {
     common::SpeedPoint speed_point;
     heuristic_speed_data_.EvaluateByTime(time_stamp, &speed_point);
-    float ref_s = speed_point.s() + init_sl_point_.s();
+    double ref_s = speed_point.s() + init_sl_point_.s();
     if (ref_s < start_s) {
       continue;
     }
@@ -192,9 +192,9 @@ ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
       break;
     }
 
-    const float s = ref_s - start_s;  // s on spline curve
-    const float l = curve.Evaluate(0, s);
-    const float dl = curve.Evaluate(1, s);
+    const double s = ref_s - start_s;  // s on spline curve
+    const double l = curve.Evaluate(0, s);
+    const double dl = curve.Evaluate(1, s);
 
     const common::SLPoint sl = common::util::MakeSLPoint(ref_s, l);
     const Box2d ego_box = GetBoxFromSLPoint(sl, dl);
@@ -203,23 +203,23 @@ ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
           GetCostBetweenObsBoxes(ego_box, obstacle_trajectory.at(index));
     }
   }
-  constexpr float kDynamicObsWeight = 1e-6;
+  constexpr double kDynamicObsWeight = 1e-6;
   obstacle_cost.safety_cost *=
       (config_.eval_time_interval() * kDynamicObsWeight);
   return obstacle_cost;
 }
 
 ComparableCost TrajectoryCost::GetCostFromObsSL(
-    const float adc_s, const float adc_l, const SLBoundary &obs_sl_boundary) {
+    const double adc_s, const double adc_l, const SLBoundary &obs_sl_boundary) {
   const auto &vehicle_param =
       common::VehicleConfigHelper::Instance()->GetConfig().vehicle_param();
 
   ComparableCost obstacle_cost;
 
-  const float adc_front_s = adc_s + vehicle_param.front_edge_to_center();
-  const float adc_end_s = adc_s - vehicle_param.back_edge_to_center();
-  const float adc_left_l = adc_l + vehicle_param.left_edge_to_center();
-  const float adc_right_l = adc_l - vehicle_param.right_edge_to_center();
+  const double adc_front_s = adc_s + vehicle_param.front_edge_to_center();
+  const double adc_end_s = adc_s - vehicle_param.back_edge_to_center();
+  const double adc_left_l = adc_l + vehicle_param.left_edge_to_center();
+  const double adc_right_l = adc_l - vehicle_param.right_edge_to_center();
 
   if (adc_left_l + FLAGS_lateral_ignore_buffer < obs_sl_boundary.start_l() ||
       adc_right_l - FLAGS_lateral_ignore_buffer > obs_sl_boundary.end_l()) {
@@ -242,7 +242,7 @@ ComparableCost TrajectoryCost::GetCostFromObsSL(
     return obstacle_cost;
   }
 
-  const float delta_l = std::fabs(
+  const double delta_l = std::fabs(
       adc_l - (obs_sl_boundary.start_l() + obs_sl_boundary.end_l()) / 2.0);
 
   const double kSafeDistance = 1.0;
@@ -252,7 +252,7 @@ ComparableCost TrajectoryCost::GetCostFromObsSL(
         Sigmoid(config_.obstacle_collision_distance() - delta_l);
   }
 
-  const float delta_s = std::fabs(
+  const double delta_s = std::fabs(
       adc_s - (obs_sl_boundary.start_s() + obs_sl_boundary.end_s()) / 2.0);
   obstacle_cost.safety_cost +=
       config_.obstacle_collision_cost() *
@@ -265,7 +265,7 @@ ComparableCost TrajectoryCost::GetCostBetweenObsBoxes(
     const Box2d &ego_box, const Box2d &obstacle_box) const {
   ComparableCost obstacle_cost;
 
-  const float distance = obstacle_box.DistanceTo(ego_box);
+  const double distance = obstacle_box.DistanceTo(ego_box);
   if (distance > config_.obstacle_ignore_distance()) {
     return obstacle_cost;
   }
@@ -279,15 +279,15 @@ ComparableCost TrajectoryCost::GetCostBetweenObsBoxes(
 }
 
 Box2d TrajectoryCost::GetBoxFromSLPoint(const common::SLPoint &sl,
-                                        const float dl) const {
+                                        const double dl) const {
   Vec2d xy_point;
   reference_line_->SLToXY(sl, &xy_point);
 
   ReferencePoint reference_point = reference_line_->GetReferencePoint(sl.s());
 
-  const float one_minus_kappa_r_d = 1 - reference_point.kappa() * sl.l();
-  const float delta_theta = std::atan2(dl, one_minus_kappa_r_d);
-  const float theta =
+  const double one_minus_kappa_r_d = 1 - reference_point.kappa() * sl.l();
+  const double delta_theta = std::atan2(dl, one_minus_kappa_r_d);
+  const double theta =
       common::math::NormalizeAngle(delta_theta + reference_point.heading());
   return Box2d(xy_point, theta, vehicle_param_.length(),
                vehicle_param_.width());
@@ -295,7 +295,7 @@ Box2d TrajectoryCost::GetBoxFromSLPoint(const common::SLPoint &sl,
 
 // TODO(All): optimize obstacle cost calculation time
 ComparableCost TrajectoryCost::Calculate(const QuinticPolynomialCurve1d &curve,
-                                         const float start_s, const float end_s,
+                                         const double start_s, const double end_s,
                                          const uint32_t curr_level,
                                          const uint32_t total_level) {
   ComparableCost total_cost;
